Fixes issues with multi-byte utf-8 characters
Using `JSON.stringify` to determine the `Content-Length` breaks when the data includes utf-8 multi-byte characters. The length only works in cases where the data only contains single byte characters. If any multi-byte characters are included it breaks. Superagent does the right thing, and uses `Buffer.byteLength()` if the `Content-Length` header is not specified.